abstract out Layout components
